ss-167: * rename modal to budget-modal
