+++ frontend/src/bundles/common/pages/budgets/budget-details/components/components.ts
@@ -1,6 +1,6 @@
+export { EditBudgetModal } from './budget-modal/budget-modal';
 export { BudgetProgressBar } from './budget-progress-bar/budget-progress-bar';
 export { RenderCurrency } from './currency/currency';
-export { EditBudgetModal } from './edit-budget-modal/edit-budget-modal';
 export { InfoCard } from './info-card/info-card';
 export { RenderMultiDropdown } from './multi-dropdown/multi-dropdown';
 export { RenderDate } from './one-day-calendar/one-day-calendar';
